requirements.txt
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-log==0.4.0
colorama==0.4.6
comm==0.1.2
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
entrypoints==0.4
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.30
gotrue==0.5.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
invoke==1.7.3
ipython==8.9.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
keyring==23.13.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.0.0
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
postgrest-py==0.10.4
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-gitlab==3.13.0
python-semantic-release==7.32.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.0
readme-renderer==37.3
realtime==0.0.5
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
storage3==0.3.6
streamlit==1.17.0
StrEnum==0.4.9
supabase==0.7.1
supafunc==0.2.2
toml==0.10.2
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
twine==3.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
zipp==3.12.0
